+++ docs/src/IconButton.doc.js
@@ -55,7 +55,7 @@ card(
       },
       {
         name: 'iconColor',
-        type: `"blue" | "darkGray" | "gray" | "red" | "white"`,
+        type: `"blue" | "darkGray" | "gray" | "red" | "white" | "orange"`,
         defaultValue: 'gray',
         href: 'iconColorCombinations',
       },